New PJax plugin, added Compass, sorted out layout
I wrote a plugin that allows for use of pjax -
it will only render the content of the given posts
or pagesi at the same URL as the original, just prefixed with
"/pjax/", then we can ignore the pages with a robots.txt.

Added Compass, but haven't really done anything with it
yet.

For categories, _posts/work/ will be where we can add
new posts, and there is a `work.html` page that will
exist at "work/index.html"-we can use it to list all of
the posts that have the category "work".